requirements.txt
pandas
duckdb
pyyaml
streamlit
plotly
